update cmake + update to sdl3

+++ include/Application.hpp
@@ -5,9 +5,9 @@
 #include <assert.h>
 #include <iostream>
 #include <glad/glad.h>
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 #include <imgui.h>
-#include <backends/imgui_impl_sdl.h>
+#include <backends/imgui_impl_sdl3.h>
 #include <backends/imgui_impl_opengl3.h>
 #include "Core.hpp"
 #include "Mesh.hpp"
@@ -104,8 +104,8 @@ namespace Engine
         //Mouse input
         int lastMouseX;
         int lastMouseY;
-        int mouseX;
-        int mouseY;
+        float mouseX;
+        float mouseY;
         int mouseOffsetX;
         int mouseOffsetY;
